test subscriptionCIDRsModel Singleton
