+++ apps/backoffice/.env.production
@@ -7,6 +7,3 @@ NEXT_PUBLIC_SELFCARE_TOKEN_EXCHANGE_URL=/mocks/token-exchange
 
 NEXT_PUBLIC_API_BACKEND_BASE_URL=http://localhost:3002
 NEXT_PUBLIC_API_BACKEND_MOCKING=true
-
-#Legacy Cosmos Uri
-LEGACY_COSMOSDB_URI=https://io-p-cosmos-api.documents.azure.com:443/
